% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{addFA}
\alias{addFA}
\title{Add missing FA annotations}
\usage{
addFA(msbatch, dmz = 5, faid, adducts = "M-H", mz, from, to)
}
\arguments{
\item{msbatch}{annotated msbatch.}

\item{dmz}{mz tolerance in ppm.}

\item{faid}{character vector specifying FA names (i.e. "FA(16:1)").}

\item{adducts}{character vector specifying adducts.}

\item{mz}{numeric vector specifying FA mz.}

\item{from}{numeric vector specifying the peak start.}

\item{to}{numeric vector specifying the peak end.}
}
\value{
annotated msbatch.
}
\description{
Add missing FA annotations
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
