% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{annotateFA}
\alias{annotateFA}
\title{FA annotation}
\usage{
annotateFA(msbatch, dmz = 5, rt, adducts = c("M-H"), db)
}
\arguments{
\item{msbatch}{msbatch obtained from LipidMS package.}

\item{dmz}{mz tolerance in ppm.}

\item{rt}{Optional. Numeric vector of length two specifying the rt range to
search for FA.}

\item{adducts}{character vector specifying adducts.}

\item{db}{FA database. Data frame with three columns: formula, total (number
of carbons and double bounds, i.e. "18:1") and Mass.}
}
\value{
annotated msbatch.
}
\description{
FA annotation
}
\examples{
\dontrun{
msbatch <- annotateFA(msbatch, dmz = 5)
}

}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
