% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{blankSubstraction}
\alias{blankSubstraction}
\title{substract blank samples.}
\usage{
blankSubstraction(fadata, blankgroup = "blank")
}
\arguments{
\item{fadata}{fadata.}

\item{blankgroup}{name used to define blank samples group.}
}
\value{
blank substracted fadata.
}
\description{
substract blank samples.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
