% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{changeFArt}
\alias{changeFArt}
\title{Modify rt peak limits of annotated FAs}
\usage{
changeFArt(msbatch, id, from, to)
}
\arguments{
\item{msbatch}{annotated msbatch.}

\item{id}{integer vector specifying FA ids to be modified.}

\item{from}{numeric vector specifying the peak start.}

\item{to}{numeric vector specifying the peak end.}
}
\value{
annotated msbatch.
}
\description{
Modify rt peak limits of annotated FAs
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
