% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{combAcetate}
\alias{combAcetate}
\title{acetate combinations for M+0, M+1 and M+2.}
\usage{
combAcetate(M)
}
\arguments{
\item{M}{total number of carbons for the FA.}
}
\value{
acetate combination for M carbon atoms.
}
\description{
acetate combinations for M+0, M+1 and M+2.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
