% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{correctNatAb13C}
\alias{correctNatAb13C}
\title{correct data for natural abundance of 13C using accucor algorithm.}
\usage{
correctNatAb13C(fadata, resolution = 140000, purity = 0.99)
}
\arguments{
\item{fadata}{fadata.}

\item{resolution}{resolution of the mass spectrometer.}

\item{purity}{purity of the tracer employed.}
}
\value{
corrected fadata.
}
\description{
correct data for natural abundance of 13C using accucor algorithm.
}
\references{
Su X, Lu W, Rabinowitz J (2017). “Metabolite Spectral Accuracy on 
Orbitraps.” Analytical Chemistry, 89(11), 5940-5948, PMID: 28471646, R package 
version 0.2.4 (2021), <https://doi.org/10.1021/acs.analchem.7b00396>.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
