% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dataElong}
\alias{dataElong}
\title{data list for elongation analysis.}
\usage{
dataElong(S16, D1, D2, P, E1, E2, E3, E4, E5, M, vcomb16, mcombe)
}
\arguments{
\item{S16}{fraction of newly synthesized C16 FA. If different to NULL it is
included in the data list.}

\item{D1}{positive numeric between 0 and 1 specifying the contribution of
acetate M+1.}

\item{D2}{positive numeric between 0 and 1 specifying the contribution of
acetate M+2.}

\item{P}{overdispersion parameter.}

\item{E1}{fraction of elongated C18 FA from C16. If different to NULL it is
included in the data list.}

\item{E2}{fraction of elongated C20 FA from C18. If different to NULL it is
included in the data list.}

\item{E3}{fraction of elongated C22 FA from C20. If different to NULL it is
included in the data list.}

\item{E4}{fraction of elongated C24 FA from C22. If different to NULL it is
included in the data list.}

\item{E5}{fraction of elongated C26 FA from C124. If different to NULL it is
included in the data list.}

\item{M}{total number of carbons for the FA.}

\item{vcomb16}{list of acetate combinations for C16 synthesis.}

\item{mcombe}{list of acetate combinations for each elongation step.}
}
\value{
data list for elongation analysis.
}
\description{
data list for elongation analysis.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
