% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{dataSynth}
\alias{dataSynth}
\title{data list for synthesis analysis.}
\usage{
dataSynth(D1, D2, P, M, vcomb)
}
\arguments{
\item{D1}{positive numeric between 0 and 1 specifying the contribution of
acetate M+1. If different to NULL it is included in the data list.}

\item{D2}{positive numeric between 0 and 1 specifying the contribution of
acetate M+2. If different to NULL it is included in the data list.}

\item{P}{overdispersion parameter. If different to NULL it is included in the
data list.}

\item{M}{total number of carbons for the FA.}

\item{vcomb}{list of acetate combinations for C16 synthesis.}
}
\value{
Data list for synthesis analysis.
}
\description{
data list for synthesis analysis.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
