\name{desaturationsdb}
\alias{desaturationsdb}
\docType{data}
\title{
Desaturation reactions database.
}
\description{
Desaturation reactions database.
}
\usage{data("desaturationsdb")}
\format{
  A data frame with 13 observations on the following 3 variables.
  \describe{
    \item{\code{precursor}}{character vector.}
    \item{\code{product}}{character vector.}
    \item{\code{parameter}}{parameter required to estimate desaturation.}
  }
}
\examples{
data(desaturationsdb)
}
\keyword{datasets}
