\name{examplefadata}
\alias{examplefadata}
\docType{data}
\title{
Example fadata list.
}
\description{
Example fadata list.
}
\usage{data("examplefadata")}
\format{
  A list with 4 elements.
  \describe{
    \item{\code{metadata}}{data frame with metadata information for samples.}
    \item{\code{fattyacids}}{data frame with compound name and label for each isotopologue (intensities df).}
    \item{\code{IS}}{data frame with IS intensities for each sample.}
    \item{\code{intensities}}{data frame with isotopologue intensities for each sample.}
  }
}
\examples{
data(fexamplefadata)
}
\keyword{datasets}
