% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{externalNormalization}
\alias{externalNormalization}
\title{External normalization using additional measures (i.e. protein levels).}
\usage{
externalNormalization(fadata, externalnormalization)
}
\arguments{
\item{fadata}{fadata list.}

\item{externalnormalization}{column names of metadata data frame used to
define external measures.}
}
\value{
normalised fadata by external measures.
}
\description{
External normalization using additional measures (i.e. protein levels).
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
