\name{fattyacidsdb}
\alias{fattyacidsdb}
\docType{data}
\title{
Fatty Acids database.
}
\description{
Fatty Acids database.
}
\usage{data("fattyacidsdb")}
\format{
  A data frame with 35 observations on the following 3 variables.
  \describe{
    \item{\code{formula}}{a character vector.}
    \item{\code{total}}{a character vector. Number of carbons and double bounds.}
    \item{\code{Mass}}{a numeric vector.}
  }
}
\examples{
data(fattyacidsdb)
}
\keyword{datasets}
