% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{fitSSE}
\alias{fitSSE}
\title{Fit FA isotope distribution with non-linear regression and calculate the sum
of squared estimate of errors.}
\usage{
fitSSE(formula, startpars, datanls, control, limitPhi = 0.1)
}
\arguments{
\item{formula}{formula.}

\item{startpars}{starting points for parameters to estimate.}

\item{datanls}{data list.}

\item{control}{nls control.}
}
\value{
sum of squared estimate errors.
}
\description{
Fit FA isotope distribution with non-linear regression and calculate the sum
of squared estimate of errors.
}
\references{
Daniel Padfield and Granville Matheson (2020). nls.multstart: Robust Non-Linear Regression using AIC Scores. R package version 1.2.0. <https://CRAN.R-project.org/package=nls.multstart>
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
