% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getResultsTable}
\alias{getResultsTable}
\title{Obtain the final results table from synthesis, elongation and desaturation
analysis.}
\usage{
getResultsTable(fadata, parameters = FAMetA::parameters)
}
\arguments{
\item{fadata}{fadata list.}

\item{parameters}{parameters to be estimated for each fatty acid. It can be
modified to change them or to add new fatty acids.}
}
\value{
results data frame.
}
\description{
Obtain the final results table from synthesis, elongation and desaturation
analysis.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
