% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{getSummaryTable}
\alias{getSummaryTable}
\title{Obtain a summary results table.}
\usage{
getSummaryTable(fadata, resultstable, parameters = FAMetA::parameters)
}
\arguments{
\item{fadata}{fadata list.}

\item{resultstable}{results data frame obtained with \link{getResultsTable}.}

\item{parameters}{parameters to be estimated for each fatty acid. It can be
modified to change them or to add new fatty acids.}
}
\value{
summary data frame.
}
\description{
Obtain a summary results table containing means and standard deviation for
each relevant parameter for each fatty acids by sample groups.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
