% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{gridElong}
\alias{gridElong}
\title{generate starting points grid for elongation analysis.}
\usage{
gridElong(S16 = NA, E1 = NA, E2 = NA, E3 = NA, E4 = NA, E5 = NA, startpoints)
}
\arguments{
\item{S16}{fraction of newly synthesized C16 FA. If NA it is included in
the starting points grid.}

\item{E1}{fraction of elongated C18 FA from C16. If NA it is included in
the starting points grid.}

\item{E2}{fraction of elongated C20 FA from C18. If NA it is included in
the starting points grid.}

\item{E3}{fraction of elongated C22 FA from C20. If NA it is included in
the starting points grid.}

\item{E4}{fraction of elongated C24 FA from C22. If NA it is included in
the starting points grid.}

\item{E5}{fraction of elongated C26 FA from C124. If NA it is included in
the starting points grid.}

\item{startpoints}{positive integer specifying the number of starting points
for each parameter to be estimated.}
}
\value{
Starting points grid for elongation analysis.
}
\description{
generate starting points grid for elongation analysis.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
