% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{gridSynth}
\alias{gridSynth}
\title{generate starting points grid for  synthesis analysis.}
\usage{
gridSynth(D1 = NA, D2 = NA, P = NA, S = NA, M, startpoints)
}
\arguments{
\item{D1}{positive numeric between 0 and 1 specifying the contribution of
acetate M+1. If NA it is included in the starting points grid.}

\item{D2}{positive numeric between 0 and 1 specifying the contribution of
acetate M+2. If NA it is included in the starting points grid.}

\item{P}{overdispersion parameter. If NA, it is included in the starting
points grid.}

\item{S}{fraction of newly synthesized C16 FA. If NA it is included in the
starting points grid.}

\item{M}{total number of carbons for the FA.}

\item{startpoints}{positive integer specifying the number of starting points
for each parameter to be estimated.}
}
\value{
Starting points grid for synthesis analysis.
}
\description{
generate starting points grid for synthesis analysis.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
