% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{heatmapResults}
\alias{heatmapResults}
\title{Plot results on a heatmap.}
\usage{
heatmapResults(toPlot, cols, scale = "none", breaks, nacolor = "grey")
}
\arguments{
\item{toPlot}{data frame with data to be plotted.}

\item{cols}{colors for the side bar (by group).}

\item{scale}{"none", "row" or "column".}

\item{breaks}{numeric vector with breaks for colouring. Optional.}

\item{nacolor}{color for NA values. Grey by default.}
}
\value{
heatmap.
}
\description{
Plot results on a heatmap.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
