% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{normalizeIS}
\alias{normalizeIS}
\title{Data normalization using internal stardards.}
\usage{
normalizeIS(fadata)
}
\arguments{
\item{fadata}{fadata list.}
}
\value{
normalised fadata by IS.
}
\description{
Data normalization using internal stardards.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
