\name{parameters}
\alias{parameters}
\docType{data}
\title{
Parameters for FA metabolic analysis.
}
\usage{data("parameters")}
\description{Parameters for FA metabolic analysis.}
\format{
  A data frame with 304 observations on the following 8 variables.
  \describe{
    \item{\code{FattyAcid}}{a character vector.}
    \item{\code{M}}{integer vector. Number of carbons.}
    \item{\code{S16}}{De novo synthesis. If equal to 1 it is estimated.}
    \item{\code{E1}}{a numeric vector. If equal to 1 it is estimated.}
    \item{\code{E2}}{a numeric vector. If equal to 1 it is estimated.}
    \item{\code{E3}}{a numeric vector. If equal to 1 it is estimated.}
    \item{\code{E4}}{a numeric vector. If equal to 1 it is estimated.}
    \item{\code{E5}}{a numeric vector. If equal to 1 it is estimated.}
  }
}
\examples{
data(paramters)
}
\keyword{datasets}
