% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{plotDistributions}
\alias{plotDistributions}
\title{Plot observed and predicted isotope distributions.}
\usage{
plotDistributions(results, groups, title = "")
}
\arguments{
\item{results}{results.}

\item{groups}{character vector specifying sample groups.}

\item{title}{title.}
}
\value{
plot.
}
\description{
Plot observed and predicted isotope distributions.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
