% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{plotResults}
\alias{plotResults}
\title{Plot observed and predicted isotope distributions from the results obtained.}
\usage{
plotResults(results, groups, fas)
}
\arguments{
\item{results}{results.}

\item{groups}{character vector specifying sample groups.}

\item{fas}{character vector specifying the FAs to be plotted.}
}
\value{
plot.
}
\description{
Plot observed and predicted isotope distributions from the results obtained.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
