% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{removeFA}
\alias{removeFA}
\title{Remove incorrect FA annotations}
\usage{
removeFA(msbatch, ids)
}
\arguments{
\item{msbatch}{annotated msbatch.}

\item{ids}{integer vector specifying FA ids to be removed.}
}
\value{
annotated msbatch.
}
\description{
Remove incorrect FA annotations
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
