% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{searchFAisotopes}
\alias{searchFAisotopes}
\title{Search FA isotopes}
\usage{
searchFAisotopes(msbatch, dmz = 5, coelCutoff = 0.7)
}
\arguments{
\item{msbatch}{annotated msbatch.}

\item{dmz}{mz tolerance in ppm.}

\item{coelCutoff}{coelution score threshold between parent and isotope peaks.}
}
\value{
fadata list.
}
\description{
Search FA isotopes
}
\examples{
\dontrun{
fadata <- searchFAisotopes(msbatch, dmz = 10, coelCutoff = 0.4)
}

}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
