% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{searchIS}
\alias{searchIS}
\title{Search internal stardards.}
\usage{
searchIS(msbatch, mz, rt, minRT, maxRT, dmz = 10)
}
\arguments{
\item{msbatch}{annotated msbatch.}

\item{mz}{numeric vector specifying IS mz.}

\item{rt}{numeric vector specifying IS rt.}

\item{minRT}{numeric vector specifying lower limits for IS rt.}

\item{maxRT}{numeric vector specifying upper limits for IS rt.}

\item{dmz}{mz tolerance in ppm.}
}
\value{
annotated msbatch.
}
\description{
Search internal stardards.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
