\name{ssexamplefadata}
\alias{ssexamplefadata}
\docType{data}
\title{
Toy example fadata list.
}
\description{
Toy example fadata list.
}
\usage{data("ssexamplefadata")}
\format{
  A list with 4 elements.
  \describe{
    \item{\code{metadata}}{data frame with metadata information for samples.}
    \item{\code{fattyacids}}{data frame with compound name and label for each isotopologue (intensities df).}
    \item{\code{IS}}{data frame with IS intensities for each sample.}
    \item{\code{intensities}}{data frame with isotopologue intensities for each sample.}
  }
}
\examples{
data(ssexamplefadata)
}
\keyword{datasets}
