% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{synthesisqmult}
\alias{synthesisqmult}
\title{calculate FA isotope distribution for newly synthesized FAs using the quasi
multinomial distribution.}
\usage{
synthesisqmult(D1, D2, P, S, M, vcomb)
}
\arguments{
\item{D1}{tracer contribution to M+1 acetate pool.}

\item{D2}{tracer contribution to M+2 acetate pool.}

\item{P}{overdispersion parameter. If different to 0, quasi-multinomial
distribution is obtained, while if set to 0, it is simplified to a multinomial
distribution.}

\item{S}{fraction of newly synthesized FA.}

\item{M}{total number of carbons for the FA.}

\item{vcomb}{list of acetate combinations obtained with \link{combAcetate}
function.}
}
\value{
numeric vector with the FA isotope distribution.
}
\description{
calculate FA isotope distribution for newly synthesized FAs using the quasi
multinomial distribution.
}
\author{
M Isabel Alcoriza-Balaguer <maribel_alcoriza@iislafe.es>
}
\keyword{internal}
